Update Wild VCF loader test fixtures and split to files
